feat: integrate liquid vesting page into burn waitlist library and clean up MintPage
