+++ libs/burn-waitlist/src/index.ts
@@ -1,5 +1,6 @@
 export * from './lib/waitlist-page';
 export * from './lib/mint-page';
+export * from './lib/liquid-vesting-page';
 export * from './lib/hooks';
 export * from './lib/components';
 export * from './lib/constants/waitlist-config';
